requirements.txt
# Core dependencies for Ollama Research Agent
ollama>=0.6.0
python-dotenv>=1.0.0

# PDF generation
reportlab>=4.0.0

# Database (SQLite3 is built into Python, no need to install)

# Optional: For better terminal output
colorama>=0.4.6
rich>=13.0.0